feat: add preview
