+++ src/game.py
@@ -1,7 +1,6 @@
-from pieces import Player
+from pieces import Player, Tetromino, TetrominoQueue
 from variables import *
 from random import choice
-import numpy as np
 
 class Timer:
     def __init__(self, duration, repeated = False, function=None):
@@ -30,14 +29,41 @@ class Timer:
             if self.repeated:
                 self.activate()
 
+class Preview:
+    def __init__(self, next_shapes):
+        self.display_surface = pygame.display.get_surface()
+        self.surface = pygame.Surface((SIZE_BAR_WIDTH, GAME_HEIGHT * PREVIEW_HEIGHT_SECTION))
+        self.rect = self.surface.get_rect(topright=(WIDTH - PADDING, PADDING))
+        self.next_shapes = next_shapes
+        self.shape_surfaces = {shape: pygame.image.load(os.path.join(IMAGE_DIR, f"{shape}.png")).convert_alpha() for shape in TETROMINO.keys()}
+        self.fragment_height = self.surface.get_height() / 3
+
+    def draw(self):
+        self.surface.fill(BLACK)
+        self.draw_pieces()
+        self.display_surface.blit(self.surface, self.rect)
+
+    def draw_pieces(self):
+        for i, shape in enumerate(self.next_shapes):
+            surface = self.shape_surfaces[shape]
+            x = (SIZE_BAR_WIDTH - surface.get_width()) // 2
+            y = i * self.fragment_height + (self.fragment_height - surface.get_height()) // 2
+            self.surface.blit(surface, (x, y))
+
+
+
 class Game:
     def __init__(self):
         self.win = pygame.display.get_surface()
         self.surface = pygame.Surface((GAME_WIDTH, GAME_HEIGHT))
         self.sprites = pygame.sprite.Group()
 
-        # Test
+        # Setting up tetromino
         self.new_tetromino()
+        self.next_shapes = [choice(list(TETROMINO.keys())) for _ in range(3)]
+        self.preview = Preview(self.next_shapes)
+
+        # Setting up game behaviours
         self.timers = {
             "gravity": Timer(300, True, self.move_down),
             "horizontal move": Timer(200),
@@ -48,6 +74,9 @@ class Game:
 
         # Blocked grid
         self.blocked_positions = {}
+
+        # Get next shapes
+        print(self.next_shapes)
 
     def check_full_row(self):
         for row in range(ROWS):
@@ -110,10 +139,6 @@ class Game:
         score_section = pygame.Surface((SIZE_BAR_WIDTH, GAME_HEIGHT * SCORE_HEIGHT_SECTION - PADDING))
         self.win.blit(score_section, (GAME_WIDTH + PADDING * 2, GAME_HEIGHT * PREVIEW_HEIGHT_SECTION + PADDING * 2))
 
-    def draw_prevew(self):
-        preview_section = pygame.Surface((SIZE_BAR_WIDTH, GAME_HEIGHT * PREVIEW_HEIGHT_SECTION))
-        self.win.blit(preview_section, (GAME_WIDTH + PADDING * 2, PADDING))
-
     def draw_grid(self):
         for col in range(1, COLUMNS):
             x = col * CELL_SIZE
@@ -142,98 +167,8 @@ class Game:
             
             # Draw the preview section
             self.draw_grid()
-            self.draw_prevew()
+            self.preview.draw()
             self.draw_score()
 
             self.win.blit(self.surface, (PADDING,PADDING))
             pygame.display.update()
-    
-class Block(pygame.sprite.Sprite):
-    def __init__(self, group, pos, color):
-        super().__init__(group)
-        self.image = pygame.Surface((CELL_SIZE, CELL_SIZE))
-        self.image.fill(color)
-
-        # Position
-        self.pos = pygame.Vector2(pos)
-        x = self.pos.x * CELL_SIZE
-        y = self.pos.y * CELL_SIZE
-        self.rect = self.image.get_rect(topleft=(x, y))
-
-    def update(self):
-        x = self.pos.x * CELL_SIZE
-        y = self.pos.y * CELL_SIZE
-        self.rect = self.image.get_rect(topleft=(x, y))
-
-    def horizontal_collision(self, blocked_positions, direction):
-        if not 0 <= self.pos.x + direction < COLUMNS:
-            return True
-        if (self.pos.x + direction, self.pos.y) in blocked_positions:
-            return True
-        return False
-        
-    def vertical_collision(self, blocked_positions):
-        if not self.pos.y + 1 < ROWS:
-            return True
-        if (self.pos.x, self.pos.y + 1) in blocked_positions:
-            return True
-        return False
-    
-class Tetromino:
-    def __init__(self, shape, group):
-        super().__init__()
-        self.shape = shape
-        self.block_positions = TETROMINO[shape]['shape']
-        self.color = TETROMINO[shape]['color']
-        self.blocks = [Block(group, (pos[0] + COLUMNS // 2, pos[1]), self.color) for pos in self.block_positions]
-        self.rotation = 0
-        # State
-        self.active = True
-
-    def move_down(self, blocked_positions):
-        if not self.check_vertical_collision(blocked_positions):
-            for block in self.blocks:
-                block.pos.y += 1
-        else:
-            self.active = False
-
-    def move_horizontal(self, blocked_positions, direction):
-        if not self.check_horizontal_collision(blocked_positions, direction):
-            for block in self.blocks:
-                block.pos.x += direction
-    
-    def check_horizontal_collision(self, blocked_positions, direction):
-        collision_list = [block.horizontal_collision(blocked_positions, direction) for block in self.blocks]
-        return any(collision_list)
-
-    def check_vertical_collision(self, blocked_positions):
-        collision_list = [block.vertical_collision(blocked_positions) for block in self.blocks]
-        return any(collision_list)
-
-    def check_rotation_collision(self, blocked_positions):
-        pivot_point = self.blocks[0].pos
-        for block in self.blocks:
-            testpos = pygame.Vector2(block.pos)
-            relative_pos = testpos - pivot_point
-            rotated_pos = relative_pos.rotate(90)
-            updated_pos = rotated_pos + pivot_point
-            if not 0 <= updated_pos.x < COLUMNS or not 0 <= updated_pos.y < ROWS:
-                return True
-            if (updated_pos.x, updated_pos.y) in blocked_positions:
-                return True
-        return False
-
-    def rotate(self, blocked_positions):
-        if self.shape == "O":
-            return
-        pivot_point = self.blocks[0].pos
-        # Make a copy of the blocks
-    
-        if self.check_rotation_collision(blocked_positions):
-            return
-
-        for block in self.blocks:
-            relative_pos = block.pos - pivot_point
-            rotated_pos = relative_pos.rotate(90)
-            updated_pos = rotated_pos + pivot_point
-            block.pos = updated_pos
